moved directives to its own class
