+++ TastyScript/Lang/TokenParser.cs
@@ -70,7 +70,7 @@ namespace TastyScript.Lang
         private void StartParse()
         {
             //start function is an inherit for easier override implementation down the line
-            ApplyDirectives();
+            new Directives();
             var startScope = FunctionStack.First("Start");
             var awakeScope = FunctionStack.First("Awake");
             HaltFunction = FunctionStack.First("Halt");
@@ -96,86 +96,6 @@ namespace TastyScript.Lang
 
             startScope.TryParse(null);
         }
-        private void ApplyDirectives()
-        {
-            var copylist = FunctionStack.List;
-            for (var i = 0; i < copylist.Count; i++) 
-                if (copylist[i].Directives != null)
-                    foreach (var d in copylist[i].Directives)
-                        switch (d.Key)
-                        {
-                            case ("Layer"):
-                                DirectiveLayer(d.Value, FunctionStack.List[i]);
-                                break;
-                            case ("Sealed"):
-                                DirectiveSealed(d.Value, FunctionStack.List[i]);
-                                break;
-                            case ("Omit"):
-                                DirectiveOmit(d.Value, FunctionStack.List[i]);
-                                break;
-                            default:
-                                Compiler.ExceptionListener.ThrowSilent(new ExceptionHandler($"Unknown directive [{d.Key}]"));
-                                break;
-                        }
-        }
-        private void DirectiveOmit(string val, IBaseFunction func)
-        {
-            val = val.Replace(" ", "").Replace("\n", "").Replace("\t", "").Replace("\r", "");
-            bool value = (val == "true" || val == "True") ? true : false;
-            if (value)
-            {
-                FunctionStack.Remove(func);
-            }
-        }
-        private void DirectiveSealed(string val, IBaseFunction func)
-        {
-            val = val.Replace(" ", "").Replace("\n", "").Replace("\t", "").Replace("\r", "");
-            bool value = (val == "true" || val == "True") ? true : false;
-            if (value)
-            {
-                func.SetSealed(value);
-                var samelist = FunctionStack.Where(func.Name).ToList();
-                var thisIndex = samelist.IndexOf(func);
-                if(thisIndex != 0)
-                {
-                    Compiler.ExceptionListener.Throw(
-                            $"Cannot override function [{func.Name}] because of the Sealed directive.");
-                }
-            }
-        }
-        private void DirectiveLayer(string val, IBaseFunction func)
-        {
-            int index = 0;
-            bool tryint = int.TryParse(val, out index);
-            if (!tryint)
-                Compiler.ExceptionListener.Throw($"The directive [Layer] must have a whole number as its value.");
-            if (index - 1 > FunctionStack.List.Count)
-                index = FunctionStack.List.Count - 1;
-            FunctionStack.MoveTo(func, index);
-            //apply any override to base connections needed since the stack was shuffled
-            var samelist = FunctionStack.Where(func.Name).ToList();
-            for(var i = 0; i < samelist.Count; i++)
-            {
-                var obj = samelist[i];
-                if (!obj.Override)
-                    continue;
-                if (i == samelist.Count - 1)
-                    continue;
-                IBaseFunction next = null;
-                for(var nexti = i; nexti <= samelist.Count - i; nexti++)
-                {
-                    if (samelist.ElementAtOrDefault(nexti) == null)
-                        continue;
-                    else
-                    {
-                        next = samelist[nexti];
-                        break;
-                    }
-                }
-                if (next == null)
-                    Compiler.ExceptionListener.Throw($"Unknown error applying directive [Layer] on function [{func.Name}]");
-                obj.SetBase(next);
-            }
-        }
+        
     }
 }
