+++ TastyScript/TastyScript.csproj
@@ -118,6 +118,7 @@
     <Compile Include="Lang\Containers\FunctionStack.cs" />
     <Compile Include="Lang\Containers\TokenStack.cs" />
     <Compile Include="Lang\Containers\StaticList.cs" />
+    <Compile Include="Lang\Directives.cs" />
     <Compile Include="Lang\Exceptions\ExceptionHandler.cs" />
     <Compile Include="Lang\Exceptions\IExceptionListener.cs" />
     <Compile Include="Lang\Extensions\ExtensionAnd.cs" />
@@ -208,7 +209,7 @@
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <ProjectExtensions>
     <VisualStudio>
-      <UserProperties BuildVersion_StartDate="2018/2/5" BuildVersion_UseGlobalSettings="True" BuildVersion_BuildVersioningStyle="None.None.None.DeltaBaseDateInDays" BuildVersion_UpdateAssemblyVersion="True" BuildVersion_UpdateFileVersion="True" BuildVersion_AssemblyInfoFilename="Properties\AssemblyInfo.cs" />
+      <UserProperties BuildVersion_AssemblyInfoFilename="Properties\AssemblyInfo.cs" BuildVersion_UpdateFileVersion="True" BuildVersion_UpdateAssemblyVersion="True" BuildVersion_BuildVersioningStyle="None.None.None.DeltaBaseDateInDays" BuildVersion_UseGlobalSettings="True" BuildVersion_StartDate="2018/2/5" />
     </VisualStudio>
   </ProjectExtensions>
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
